Finalizando api

+++ src/apptlink.Domain/Types/DetallePedidoType.cs
@@ -5,7 +5,7 @@ namespace apptlink.Domain.Types;
 
 public class DetallePedidoType
 {
-    [JsonPropertyName("id")] public long Id { get; set; }
+    [JsonPropertyName("id")] public int Id { get; set; }
     [JsonPropertyName("pedido_id")] public int PedidoId { get; set; }
     [JsonPropertyName("producto_id")] public int ProductoId { get; set; }
     [JsonPropertyName("cantidad")] public int Cantidad { get; set; }
